+++ include/delameta/url.h
@@ -30,8 +30,9 @@ struct fmt::formatter<Project::delameta::URL> {
 
     template <typename Ctx>
     inline auto format(const Project::delameta::URL& m, Ctx& ctx) const {
-        return fmt::format_to(ctx.out(), "{}: {{protocol: {}, host: {}, path: {}, full path: {}, queries: {}, fragment: {}}}", 
-            m.url, m.host, m.path, m.full_path, m.queries, m.fragment);
+        return fmt::format_to(ctx.out(), 
+            "\"{}\": {{protocol: \"{}\", host: \"{}\", path: \"{}\", full path: \"{}\", queries: {}, fragment: {}}}", 
+            m.url, m.protocol, m.host, m.path, m.full_path, m.queries, m.fragment);
     }
 };