+++ test/url.cpp
@@ -1,7 +1,4 @@
-#include <boost/preprocessor.hpp>
-#include <fmt/format.h>
 #include <delameta/url.h>
-#include <delameta/modbus/api.h>
 #include <gtest/gtest.h>
 
 using namespace Project;